Add support for multiple Handler backends
This moves the built-in signal handling into a pluggable Handler backends
which allows endless to support Windows which doesn't support signals.

The default Handler on none Windows OS's is a SignalHandler so still
maintaining drop in replacement of http.ListenAndServe methods.

Server now exports a Restart() method which is the heart of endless.

SignalHandler can hook any signal handle however only the following
are used internally:
* SIGHUP: Call Restart()
* SIGUSR2: Call Terminate(0)
* SIGINT & SIGTERM: Call Shutdown()

Hooked signals now require a function of type SignalHook which takes
the triggering os.Signal along with the Server. A Hook can prevent
further processing of the signal by returning false.

Also:
* Renamed HammerTime to TerminateTimeout
* Made state management internal using a type for compile time validation.
* Removed initialisation of variable to their default values (go vet)
* Documented methods.
* Cleaned up comment formatting for easier reading.
* Added TODO's where potential issues exist.
* Made use of http.Server ErrorLog for all logging if not nil otherwise standard logger.
* Made TerminateTimeout configurable on each server.
* Removed commented out debugging.
* Renamed type endlessServer -> Server so its exported and idiomaticly named.
* Renamed type and var endlessListener -> Listener so they exported and idomaticly named.
* Renamed command line flags to avoid conflict with existing application flags.
* Fixed err var shadowing.
* Fix syscall.Kill calls on Windows.
* Remove isChild from Server as its not needed, use the global instead.
* Add state transition validation and error checking.
* Exported GetState
* Use a done channel to fast bypass a Terminate request.
* Corrected comment on keep-alive disable.
* Use a safe type cast instead of type switch in Restart.