Add initial endless tests
Add initial tests to validate endless operation.

This identified the face the use of panic from excessive wg.Done() calls in Terminate wasn't actually working as the recover only recovers the current goroutine.

What was happening was when a handler went to exit normal it would panic and bring the entire server down.

Instead we now store each Conn and force close them on terminate; this won't prevent the handlers from continuing to attempt to process but does allow the server to cleanly exit.

Also:
* Add StateFailed
* Remove unneeded wg and lock init in Server{}
* Hide output under Debug flag to support tests properly.
* Removed disable keep-alives as it only happened for new connections and we close the socket on the next line.
* Remove stopped from Listener as it was racey.
* Lock conn.Close and use nil listener as the closed flag preventing missing wg.Done if Close returned an unexpected error.
* Add listener method which returns the current *Listener
* Fixed comment typo's