Refactor to a single server manager
Previously a server had Restart, Shutdown, Terminate etc methods but these are operations that only make sense globally.

In order to fix this we now have a package level Manager which is responsible for these operations.

As a side effect of this Handlers now operate on the Manager and not Servers directly.

Also:
* Allow none tcp listeners
* Add restart tests
* Fix restarts with multiple servers. Prior to this rework servers whould be shutdown after the first server had started.
* Fix windows support by removing reliance on SIGTERM
* Switch to ENV based socket ordering, which prevents issues with external cmd flags.
* Fix some error checking in tests.